requirements.txt
fastapi
uvicorn
tweepy
python-dotenv
pydantic
facebook-sdk
requests
nltk
